pyproject.toml
[tool.poetry]
name = "2023-hackathon"
version = "0.1.0"
description = ""
authors = ["Bing Wang <bing.wang@flatiron.com>"]
readme = "README.md"
packages = [{include = "2023_hackathon"}]

[tool.poetry.dependencies]
python = "^3.10"
redis = "^4.6.0"
numpy = "^1.25.1"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
azure-identity = "^1.13.0"
pypdf2 = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
